Set-up
Create a 2D platformer controller in Unity that can reliably handle slopes and moving platforms.

----------------------------------------------------------------------------------------------------------------
Basic - Using two scripts:

1- Player
	Responsible to receiving input

2- Controller
	Move the player
	Fire out rays horizontally and vertically in the direction that the player is moving (overly constraining to collisions)

OBSERVATION: Rays aren't actually being fired from the exact edge of the player's bounds but rather from a small width inside of them witch  refer to as the skin width. The reason for this is so that when the player is say resting flush on the ground we still have a small amount of space within which we can effectively fire the Rays.

----------------------------------------------------------------------------------------------------------------

Final point:

Q: Why we bother with all of this Raycast stuff when we could just use a rigidbody to handle the collision detection for us?
A: The raycast controller handles slopes and moving platforms far more consistentily than the rigidbody and also there it gives one sort of far greater fine-tuned control over the player movement.